requirements.txt
langchain
langchain-community
faiss-cpu
langchain-google-genai
google-generativeai
fastapi
streamlit
python-multipart